Many Issues Fixed

+++ RedGroup3DMaze/src/Selection.java
@@ -15,7 +15,6 @@ public class Selection extends JPanel implements ActionListener{
 	int difficulty; 
 	JFrame frame; 
 	JPanel panel;
-	//JLabel title, logo, background;
 	JLabel background;
 	JButton easyButt, mediumButt, hardButt; 
 	private boolean signal = false;
@@ -36,14 +35,7 @@ public class Selection extends JPanel implements ActionListener{
 		hardButt = new JButton(new ImageIcon(new ImageIcon(getClass().getClassLoader().
 				getResource("hard.png")).getImage().
 				getScaledInstance(400, 100, java.awt.Image.SCALE_SMOOTH)));
-		//logo = new JLabel(new ImageIcon(getClass().getClassLoader().getResource("MazeLogo.png")));
-		//title = new JLabel(new ImageIcon(getClass().getClassLoader().getResource("Welcome.png")));	
 	}
-	
-//	public void paintComponent(Graphics g) {
-//		super.paintComponent(g);
-//		g.drawImage(new ImageIcon(ImageIO.read(getClass().getClassLoader().getResource("selBackground.png"))), 0, 0, );
-//	}
 	
 	public void display() {
 		background.setIcon(new ImageIcon(new ImageIcon(getClass().getClassLoader().
@@ -51,7 +43,6 @@ public class Selection extends JPanel implements ActionListener{
 				getScaledInstance((int)(frame.getWidth()*0.99), (int)(frame.getHeight()*0.954), 
 						java.awt.Image.SCALE_SMOOTH)));
 	    background.setBounds(-5, -17, frame.getWidth(), frame.getHeight());
-//	    background.setOpaque(true);
 		
 	    easyButt.setActionCommand("Easy");
 	    easyButt.addActionListener(this);
@@ -72,46 +63,31 @@ public class Selection extends JPanel implements ActionListener{
 	    hardButt.setContentAreaFilled(false);
 	    hardButt.setBorderPainted(false);
 	    
-	    int panelWidth = panel.getSize().width;
-	    int panelHeight = panel.getSize().height;
+	    double panelWidth = panel.getSize().width;
+	    double panelHeight = panel.getSize().height;
 	    
 	    easyButt.setIcon(new ImageIcon(new ImageIcon(getClass().getClassLoader().
 				getResource("easy.png")).getImage().
-				getScaledInstance((int)(300 * panelWidth/1000), (int)(75* panelHeight/750), 
+				getScaledInstance((int)(300.0 * panelWidth/1000.0), (int)(75.0* panelHeight/750.0), 
 						java.awt.Image.SCALE_SMOOTH)));
 	    mediumButt.setIcon(new ImageIcon(new ImageIcon(getClass().getClassLoader().
 				getResource("medium.png")).getImage().
-				getScaledInstance((int)(300 * panelWidth/1000), (int)(75* panelHeight/750), 
+				getScaledInstance((int)(300.0 * panelWidth/1000.0), (int)(75.0* panelHeight/750.0), 
 						java.awt.Image.SCALE_SMOOTH)));
 	    hardButt.setIcon(new ImageIcon(new ImageIcon(getClass().getClassLoader().
 				getResource("hard.png")).getImage().
-				getScaledInstance((int)(300 * panelWidth/1000), (int)(75* panelHeight/750), 
+				getScaledInstance((int)(300.0 * panelWidth/1000.0), (int)(75.0* panelHeight/750.0), 
 						java.awt.Image.SCALE_SMOOTH)));
 	    
 	    easyButt.setBounds((int)(panelWidth*.07), (int)(panelHeight*0.333),(int)(400*panelWidth/1000),(int)(110*panelHeight/750));
-	    //easyButt.setFont(new Font("Serif", Font.PLAIN, (int)(panelWidth*.03)));
 	    mediumButt.setBounds((int)(panelWidth*.07), (int)(panelHeight*0.5), (int)(400*panelWidth/1000),(int)(110*panelHeight/750));
-	    //mediumButt.setFont(new Font("Serif", Font.PLAIN, (int)(panelWidth*.03)));
 	    hardButt.setBounds((int)(panelWidth*.07), (int)(panelHeight*0.667), (int)(400*panelWidth/1000),(int)(110*panelHeight/750));
-	    //hardButt.setFont(new Font("Serif", Font.PLAIN, (int)(panelWidth*.03)));
-	    	    
-	    //title.setBounds((int)(panelWidth*0.01), (int)(panelHeight*.06), (int)(panelWidth*1), (int)(panelHeight*.13));
-	    	    
-	    //logo.setBounds((int)(panelWidth*0.5), (int)(panelHeight*.267), (int)(panelWidth*.4), (int)(panelHeight*.53));
-	    
-//	    try {
-//            frame.setContentPane(new JLabel(new ImageIcon(ImageIO.read(getClass().getClassLoader().getResource("selBackground.png")))));
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
 	    
 	    
 	    panel.add(easyButt);
 	    panel.add(mediumButt);
 	    panel.add(hardButt);
 	    panel.add(background);
-	    //panel.add(logo);
-	    //panel.add(title);
 	}
 	
 	public void hide(JFrame frame, JPanel panel) {
@@ -119,8 +95,6 @@ public class Selection extends JPanel implements ActionListener{
 		panel.remove(easyButt);
 		panel.remove(mediumButt);
 		panel.remove(hardButt);
-		//panel.remove(title);
-		//panel.remove(logo);
 	}
 
 //handles setting the difficulty of the maze
